Update jquery.mobile.datepicker.js
When calling $("selector").date() on a inline datepicker, the datepicker is not properly initialized, so when calling any method (exaple: $("selector").date("getDate")) causes the error: "cannot access datepicker of undefined". This is happening because in the line "this.baseWidget = ( !this.options.inline )? this.element: this.calendar;" the call to "this.calendar" is accessing the wrong scope. Changing it to "this.baseWidget = ( !this.options.inline )? this.element: calendar;" solved the problem for me.